+++ lambda/src/core/kernel.rs
