+++ lambda/src/core/mod.rs
@@ -1,5 +1,5 @@
 pub mod component;
 pub mod events;
+pub mod kernel;
 pub mod render;
-pub mod runnable;
 pub mod window;
